Update README.md
@Lasall and @Kindrey, your friends are bullies. You know this. You helped them bully me. Stop this.

Seriously, what is wrong with you?

You know this is wrong yet you come up with excuses and egg each other on to keep mistreating me.

You need to listen. What you are doing is WRONG. It is not going to get any less wrong.

Stop this insanity. Stop making excuses for abuse.